requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
imageio==2.36.0
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.4
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.4.0
six==1.16.0
soupsieve==2.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.5.0
tifffile==2024.9.20
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
